requirements.txt
google-adk
litellm
loguru
exa_py
biopython
tooluniverse